TuringABC

Documentation for TuringABC.

TuringABC.ABC

TuringABC.make_joint_model

TuringABC.observations

TuringABC.sample_from_joint

TuringABC.split_latent_data

TuringABC.split_latent_data

TuringABC.statistic_distance

TuringABC.statistic_distance

TuringABC.ABC — Type

ABC <: AbstractMCMC.AbstractSampler

Approximate Bayesian Computation (ABC) sampler.

TuringABC.make_joint_model — Method

make_joint_model(sampler::ABC, model::DynamicPPL.Model)

Return a model with observations now also considered random variables.

TuringABC.observations — Method

observations(sampler::ABC, model::DynamicPPL.Model)

Return the observations in model.

TuringABC.sample_from_joint — Method

sample_from_joint(rng::Random.AbstractRNG, sampler::ABC, model::DynamicPPL.Model)

Sample from the joint model.

Defaults to rand(rng, OrderedDict, make_joint_model(sampler, model)).

See also: make_joint_model.

TuringABC.split_latent_data — Method

split_latent_data(sampler::ABC, model::DynamicPPL.Model, x)

Return a tuple with first element being variables and second being data.

TuringABC.split_latent_data — Method

split_latent_data(d::OrderedDict, data_variable)

Return a tuple with first element being variables and second being data.

TuringABC.statistic_distance — Method

statistic_distance(sampler::ABC, data_true, data_candidate)

Return the distance between the statistics of data_true and data_candidate.

TuringABC.statistic_distance — Method

statistic_distance(sampler::ABC, model::DynamicPPL.Model, data_candidate)

Return the distance between observations in model and data_candidate.